pyproject.toml
[project]
name = "lean4evaluation"
version = "0.1.0"
description = "LLM Formal Mathematics Evaluation Framework for Lean 4"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "tenacity>=8.0.0",
    "tqdm>=4.60.0",
]
